backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
sse-starlette>=2.1.0
PyJWT>=2.9.0

# Logging
structlog>=24.4.0

# External Services
supabase>=2.10.0
anthropic>=0.40.0
deepgram-sdk>=3.8.0
resend>=2.5.0

# Type checking
mypy>=1.13.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
